Code for WDIKT test
Here is the basic code I wrote for the texted based game, Why Do I Know This? It is full of random questions and silly banter by your computer. Right now it only works on Mac, but future versions will be everywhere. Run it from the command line, typing in 
                                        python WDIKTtest.py YourName

(You will of course put your name in the place of  YourName, I just did it like that to say, no spaces, please.)